+++ plugins/providers/virtualbox/driver/base.rb
@@ -153,6 +153,11 @@ module VagrantPlugins
         def import(ovf)
         end
 
+        # Returns the maximum number of network adapters.
+        def max_network_adapters
+          8
+        end
+
         # Returns a list of forwarded ports for a VM.
         #
         # @param [String] uuid UUID of the VM to read from, or `nil` if this